Break out utils in mulit_ne/mls
