+++ packages/sdk/src/tests/multi_ne/mls/dms.test.ts
@@ -3,9 +3,9 @@
  */
 
 import { Client } from '../../../client'
-import { getChannelMessagePayload, makeTestClient } from '../../testUtils'
+import { makeTestClient } from '../../testUtils'
 import { MLS_ALGORITHM } from '../../../mls'
-import { StreamTimelineEvent } from '../../../types'
+import { checkTimelineContainsAll } from './utils'
 
 const clients: Client[] = []
 
@@ -97,36 +97,3 @@ describe('dmsMlsTests', () => {
             .toBe(true)
     }, 10_000)
 })
-
-function getPayloadRemoteEvent(event: StreamTimelineEvent): string | undefined {
-    if (event.decryptedContent?.kind === 'channelMessage') {
-        return getChannelMessagePayload(event.decryptedContent.content)
-    }
-    return undefined
-}
-
-function getPayloadLocalEvent(event: StreamTimelineEvent): string | undefined {
-    if (event.localEvent?.channelMessage) {
-        return getChannelMessagePayload(event.localEvent.channelMessage)
-    }
-    return undefined
-}
-
-function getPayload(event: StreamTimelineEvent): string | undefined {
-    const payload = getPayloadRemoteEvent(event)
-    if (payload) {
-        return payload
-    }
-    return getPayloadLocalEvent(event)
-}
-
-function checkTimelineContainsAll(messages: string[], timeline: StreamTimelineEvent[]): boolean {
-    const checks = new Set(messages)
-    for (const event of timeline) {
-        const payload = getPayload(event)
-        if (payload) {
-            checks.delete(payload)
-        }
-    }
-    return checks.size === 0
-}
